+++ src/components/cart/page/CartItem.jsx
@@ -4,36 +4,39 @@ import React, { useCallback, useState } from "react";
 import Counter from "../Counter";
 import { Link } from "react-router-dom";
 import { cartActions } from "../../../store/cart-slice";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
 
 const CartItem = ({ className, cartItem, deleting }) => {
   const [loading, setLoading] = useState(false);
   const [deleteLoading, setDeleteLoading] = useState(false);
+  const authedUser = useSelector((state) => state.auth.user);
   const dispatch = useDispatch();
 
   const handleRemoveFromCart = useCallback(
     (e) => {
       e?.preventDefault();
-      setDeleteLoading(true);
-      fetch(`http://localhost:8000/cart/${cartItem.id}`, {
-        method: "DELETE",
-        body: JSON.stringify(cartItem),
-      })
-        .then((res) => {
-          if (!res.ok) throw new Error();
-          return res.json();
+      if (authedUser) {
+        setDeleteLoading(true);
+        fetch(`http://localhost:8000/cart/${cartItem.id}`, {
+          method: "DELETE",
+          body: JSON.stringify(cartItem),
         })
-        .then((data) => {
-          setDeleteLoading(false);
-          dispatch(cartActions.removeItem(data.id));
-        })
-        .catch(() => {
-          setDeleteLoading(false);
-          toast.error("Unable to remove this item :(");
-        });
+          .then((res) => {
+            if (!res.ok) throw new Error();
+            return res.json();
+          })
+          .then((data) => {
+            setDeleteLoading(false);
+            dispatch(cartActions.removeItem(data.id));
+          })
+          .catch(() => {
+            setDeleteLoading(false);
+            toast.error("Unable to remove this item :(");
+          });
+      } else dispatch(cartActions.removeItem(cartItem.id));
     },
-    [dispatch, cartItem]
+    [dispatch, cartItem, authedUser]
   );
 
   const handleChangeValue = useCallback(
@@ -42,30 +45,38 @@ const CartItem = ({ className, cartItem, deleting }) => {
         handleRemoveFromCart();
         return;
       }
-      setLoading(true);
-      fetch(`http://localhost:8000/cart/${cartItem.id}`, {
-        method: "PUT",
-        body: JSON.stringify({
-          ...cartItem,
-          amount: newValue,
-        }),
-      })
-        .then((res) => {
-          if (!res.ok) throw new Error();
-          return res.json();
+      if (authedUser) {
+        setLoading(true);
+        fetch(`http://localhost:8000/cart/${cartItem.id}`, {
+          method: "PUT",
+          body: JSON.stringify({
+            ...cartItem,
+            amount: newValue,
+          }),
         })
-        .then((data) => {
-          setLoading(false);
-          dispatch(
-            cartActions.changeAmount({ id: data.id, newAmount: data.amount })
-          );
-        })
-        .catch(() => {
-          setLoading(false);
-          toast.error("Error changing the amount :(");
-        });
+          .then((res) => {
+            if (!res.ok) throw new Error();
+            return res.json();
+          })
+          .then((data) => {
+            setLoading(false);
+            dispatch(
+              cartActions.changeAmount({ id: data.id, newAmount: data.amount })
+            );
+          })
+          .catch(() => {
+            setLoading(false);
+            toast.error("Error changing the amount :(");
+          });
+      } else
+        dispatch(
+          cartActions.changeAmount({
+            id: cartItem.id,
+            newAmount: newValue,
+          })
+        );
     },
-    [dispatch, cartItem, handleRemoveFromCart]
+    [dispatch, cartItem, handleRemoveFromCart, authedUser]
   );
 
   return (
